chore(ui): dashboard edit preview to respect set layouts
