+++ ui/src/components/dashboard/Dashboard.vue
@@ -24,38 +24,7 @@
         <small>{{ dashboard.description }}</small>
     </section>
 
-    <section id="charts">
-        <el-row :gutter="8">
-            <el-col
-                v-for="(chart, index) in charts"
-                :key="`${chart.id}__${index}`"
-                :xs="24"
-                :sm="(chart.chartOptions?.width || 6) * 4"
-                :md="(chart.chartOptions?.width || 6) * 2"
-            >
-                <div class="d-flex flex-column">
-                    <p class="m-0">
-                        <span class="fs-6 fw-bold">{{ labels.title }}</span>
-                        <template v-if="labels?.description">
-                            <br>
-                            <small class="fw-light">
-                                {{ labels.description }}
-                            </small>
-                        </template>
-                    </p>
-
-                    <div class="flex-grow-1 mt-2">
-                        <component
-                            :is="TYPES[chart.type]"
-                            :default="route.params.id === 'default'"
-                            :source="chart.content"
-                            :chart
-                        />
-                    </div>
-                </div>
-            </el-col>
-        </el-row>
-    </section>
+    <ChartsSection :charts />
 </template>
 
 <script setup>
@@ -63,6 +32,7 @@
 
     import Header from "./components/Header.vue";
     import KestraFilter from "../filter/KestraFilter.vue";
+    import ChartsSection from "./components/ChartsSection.vue";
 
     import {useRoute, useRouter} from "vue-router";
     const router = useRouter();
@@ -73,22 +43,6 @@
 
     import {useI18n} from "vue-i18n";
     const {t} = useI18n({useScope: "global"});
-
-    import TimeSeries from "./components/charts/custom/TimeSeries.vue";
-    import Bar from "./components/charts/custom/Bar.vue";
-    import Markdown from "../layout/Markdown.vue";
-    import Table from "./components/tables/custom/Table.vue";
-    import Pie from "./components/charts/custom/Pie.vue";
-    import KPI from "./components/charts/custom/KPI.vue";
-
-    const TYPES = {
-        "io.kestra.plugin.core.dashboard.chart.TimeSeries": TimeSeries,
-        "io.kestra.plugin.core.dashboard.chart.Bar": Bar,
-        "io.kestra.plugin.core.dashboard.chart.Markdown": Markdown,
-        "io.kestra.plugin.core.dashboard.chart.Table": Table,
-        "io.kestra.plugin.core.dashboard.chart.Pie": Pie,
-        "io.kestra.plugin;core.dashboard.chart.KPI": KPI,
-    };
 
     const props = defineProps({
         embed: {type: Boolean, default: false},
@@ -103,11 +57,6 @@
 
     const dashboard = ref({});
     const charts = ref([]);
-
-    const labels = (chart) => ({
-        title: chart?.chartOptions?.displayName ?? chart?.id,
-        description: chart?.chartOptions?.description,
-    });
 
     const loadCharts = async (allCharts) => {
         charts.value = [];
@@ -152,21 +101,5 @@ section#description {
     margin: 0 0.25rem;
     padding: 0 2rem 1rem;
     color: var(--ks-content-secondary);
-}
-
-section#charts {
-    padding: 0 2rem 1rem;
-
-    & .el-row .el-col {
-        margin-bottom: 0.5rem;
-
-        & > div {
-            height: 100%;
-            padding: 1.5rem;
-            background: var(--ks-background-card);
-            border: 1px solid var(--ks-border-primary);
-            border-radius: $border-radius;
-        }
-    }
 }
 </style>
